address review comments.
